faster classloading
- split ClassPathCaches's bloomfilter into bloomfilter per Loader
- store bloomfilter on disk directly in classpath.index: 20x less disk space, faster startup

+++ platform/util/src/com/intellij/util/lang/FileLoader.java
@@ -18,6 +18,8 @@ package com.intellij.util.lang;
 import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.util.SmartList;
+import com.intellij.util.io.DataInputOutputUtil;
+import com.intellij.util.io.DataOutputStream;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -39,8 +41,8 @@ class FileLoader extends Loader {
     myConfiguration = configuration;
   }
 
-  private void buildPackageCache(final File dir, ClasspathCache.LoaderData loaderData) {
-    loaderData.addResourceEntry(getRelativeResourcePath(dir));
+  private void buildPackageCache(final File dir, ClasspathCache.LoaderDataBuilder context) {
+    context.addResourcePackageFromName(getRelativeResourcePath(dir));
 
     final File[] files = dir.listFiles();
     if (files == null) {
@@ -48,18 +50,19 @@ class FileLoader extends Loader {
     }
 
     boolean containsClasses = false;
+    
     for (File file : files) {
       final boolean isClass = file.getPath().endsWith(UrlClassLoader.CLASS_EXTENSION);
       if (isClass) {
         if (!containsClasses) {
-          loaderData.addResourceEntry(getRelativeResourcePath(file));
+          context.addClassPackageFromName(getRelativeResourcePath(file));
           containsClasses = true;
         }
-        loaderData.addNameEntry(file.getName());
+        context.addPossiblyDuplicateNameEntry(file.getName());
       }
       else {
-        loaderData.addNameEntry(file.getName());
-        buildPackageCache(file, loaderData);
+        context.addPossiblyDuplicateNameEntry(file.getName());
+        buildPackageCache(file, context);
       }
     }
   }
@@ -178,67 +181,62 @@ class FileLoader extends Loader {
   private static final AtomicLong totalReading = new AtomicLong();
 
   private static final Boolean doFsActivityLogging = false;
+  private static final int ourVersion = 1;
 
   private ClasspathCache.LoaderData tryReadFromIndex() {
     if (!myConfiguration.myCanHavePersistentIndex) return null;
     long started = System.nanoTime();
-    ClasspathCache.LoaderData loaderData = new ClasspathCache.LoaderData();
+    
     File index = getIndexFileFile();
 
-    BufferedReader reader = null;
+    DataInputStream reader = null;
+    boolean isOk = false;
+    
     try {
-      reader = new BufferedReader(new FileReader(index));
-      readList(reader, loaderData.getResourcePaths());
-      readList(reader, loaderData.getNames());
-
-      return loaderData;
-    } catch (Exception ex) {
-      if (!(ex instanceof FileNotFoundException)) index.delete();
-    }
+      reader = new DataInputStream(new BufferedInputStream(new FileInputStream(index)));
+      if (DataInputOutputUtil.readINT(reader) == ourVersion) {
+        ClasspathCache.LoaderData loaderData = new ClasspathCache.LoaderData(reader);
+        isOk = true;
+        return loaderData;
+      }
+    } catch (FileNotFoundException ex) {
+      isOk = true;
+    } catch (IOException ignore) {}
     finally {
       if (reader != null) {
         try {
           reader.close();
         } catch (IOException ignore) {}
       }
+      if (!isOk) index.delete();
       totalReading.addAndGet(System.nanoTime() - started);
     }
 
     return null;
   }
 
-  private static void readList(BufferedReader reader, List<? super String> paths) throws IOException {
-    String line = reader.readLine();
-    int numberOfElements = Integer.parseInt(line);
-    for(int i = 0; i < numberOfElements; ++i) paths.add(reader.readLine());
-  }
-
   private void trySaveToIndex(ClasspathCache.LoaderData data) {
     if (!myConfiguration.myCanHavePersistentIndex) return;
     long started = System.nanoTime();
     File index = getIndexFileFile();
-    BufferedWriter writer = null;
+    DataOutput writer = null;
+    boolean isOk = false;
 
     try {
-      writer = new BufferedWriter(new FileWriter(index));
-      writeList(writer, data.getResourcePaths());
-      writeList(writer, data.getNames());
-    } catch (IOException ex) {
-      index.delete();
-    }
+      writer = new DataOutputStream(new BufferedOutputStream(new FileOutputStream(index)));
+      DataInputOutputUtil.writeINT(writer, ourVersion);
+      data.save(writer);
+      isOk = true;
+    } catch (IOException ignore) {}
     finally {
       if (writer != null) {
         try {
-          writer.close();
+          ((OutputStream)writer).close();
         } catch (IOException ignore) {}
       }
+      if (!isOk) index.delete();
       totalSaving.addAndGet(System.nanoTime() - started);
     }
-  }
-
-  private static void writeList(BufferedWriter writer, List<String> paths) throws IOException {
-    writer.append(Integer.toString(paths.size())).append('\n');
-    for(String s: paths) writer.append(s).append('\n');
   }
 
   @NotNull
@@ -248,14 +246,13 @@ class FileLoader extends Loader {
 
   @NotNull
   @Override
-  public ClasspathCache.LoaderData buildData() throws IOException {
-    ClasspathCache.LoaderData fromIndex = tryReadFromIndex();
-    final ClasspathCache.LoaderData loaderData = fromIndex != null ? fromIndex : new ClasspathCache.LoaderData();
+  public ClasspathCache.LoaderData buildData() {
+    ClasspathCache.LoaderData loaderData = tryReadFromIndex();
 
     final int nsMsFactor = 1000000;
     int currentLoaders = totalLoaders.incrementAndGet();
-    long currentScanning;
-    if (fromIndex == null) {
+    long currentScanningTime;
+    if (loaderData == null) {
       long started = System.nanoTime();
 /*    // todo code below uses java 7 api, uncomment once we are done with Java 6
       if (SystemInfo.isJavaVersionAtLeast("1.7") && !SystemProperties.getBooleanProperty("idea.no.nio.class.scanning", false) && false) {
@@ -302,23 +299,22 @@ class FileLoader extends Loader {
           }
         });
       } else {  */
-        buildPackageCache(myRootDir, loaderData);
+      ClasspathCache.LoaderDataBuilder loaderDataBuilder = new ClasspathCache.LoaderDataBuilder();
+      buildPackageCache(myRootDir, loaderDataBuilder);
+      loaderData = loaderDataBuilder.build();
       /* } */
       final long doneNanos = System.nanoTime() - started;
-      currentScanning = totalScanning.addAndGet(doneNanos);
+      currentScanningTime = totalScanning.addAndGet(doneNanos);
       if (doFsActivityLogging) {
         System.out.println("Scanned: " + myRootDirAbsolutePath + " for " + (doneNanos / nsMsFactor) + "ms");
       }
       trySaveToIndex(loaderData);
     } else {
-      currentScanning = totalScanning.get();
+      currentScanningTime = totalScanning.get();
     }
 
-    loaderData.addResourceEntry("foo.class");
-    loaderData.addResourceEntry("bar.properties");
-
     if (doFsActivityLogging) {
-      System.out.println("Scanning: " + (currentScanning / nsMsFactor) + "ms, saving: " + (totalSaving.get() / nsMsFactor) +
+      System.out.println("Scanning: " + (currentScanningTime / nsMsFactor) + "ms, saving: " + (totalSaving.get() / nsMsFactor) +
                          "ms, loading:" + (totalReading.get() / nsMsFactor) + "ms for " + currentLoaders + " loaders");
     }
